Support overloading of create_mirror() for custom scalar types
These changes make some small modifications to allow Kokkos
specializations for custom scalar types (e.g., Sacado and Stokhos) to
overload Kokkos::create_mirror():

* Add enable_if logic to create_mirror() so that the Kokkos
  implementations are picked up only for the default View specialization.
  This is for both View's and DynRankView's.
* Move create_mirror() and related functions to Kokkos_CopyViews.hpp, so
  that Kokkos_View.hpp can be included in specializations.  This solves
  an order of include problem for Sacado/Stokhos which need to overload
  create_mirror() before the implementation of create_mirror_view() and
  without requiring numerous forward declarations.
* Remove include of Kokkos_Core.hpp from Kokkos_StaticCrsGraph.hpp and
  just include the files it uses.  This is needed by Stokhos for the
  same order of include problem.
