+++ containers/src/Kokkos_StaticCrsGraph.hpp
@@ -47,7 +47,8 @@
 #include <string>
 #include <vector>
 
-#include <Kokkos_Core.hpp>
+#include <Kokkos_View.hpp>
+#include <Kokkos_Parallel_Reduce.hpp>
 
 namespace Kokkos {
